Honour `choice` tag on positional arguments
If a positional argument has a non-slice type (i.e., if it represents a
single argument rather than a collection of remaining arguments), allow
the use of the `choice` tag to restrict the set of permitted values, as
is already possible with options. Additionally, provide choices for
positional arguments as command line completions in the appropriate
context.